+++ tests/testthat/test-utils.R
@@ -87,4 +87,9 @@ with_mock_dir("ref_table",simplify= F,{
   })
 })
 
-
+if(length(list.files(tools::R_user_dir('comtradr', which = 'cache')))==0){
+  fs::dir_delete(tools::R_user_dir('comtradr', which = 'cache'))
+}
+if(length(list.files(tools::R_user_dir('comtradr_bulk', which = 'cache')))==0){
+  fs::dir_delete(tools::R_user_dir('comtradr_bulk', which = 'cache'))
+}